Refactored Rules import/export step.
