+++ src/Orchard.Web/Modules/Orchard.Rules/Orchard.Rules.csproj
@@ -89,9 +89,8 @@
     <Content Include="Scripts\Web.config" />
     <Content Include="Styles\Web.config" />
     <Compile Include="Handlers\ContentHandler.cs" />
-    <Compile Include="ImportExport\RulesExportEventHandler.cs" />
-    <Compile Include="ImportExport\RulesCustomExportStep.cs" />
-    <Compile Include="ImportExport\RulesRecipeHandler.cs" />
+    <Compile Include="Recipes\Builders\RulesStep.cs" />
+    <Compile Include="Recipes\Executors\RulesStep.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Content Include="Module.txt" />
     <Content Include="Web.config">
